refactor(plugin-manager): Move mappings to separate file
