+++ components/plugin-manager/backend/src/main/kotlin/PluginService.kt
@@ -19,8 +19,6 @@
 
 package org.eclipse.apoapsis.ortserver.components.pluginmanager
 
-import org.eclipse.apoapsis.ortserver.components.pluginmanager.routes.mapToApi
-
 import org.jetbrains.exposed.sql.Database
 import org.jetbrains.exposed.sql.and
 import org.jetbrains.exposed.sql.selectAll
