+++ components/plugin-manager/backend/src/test/kotlin/routes/UpdateTemplateOptionsIntegrationTest.kt
@@ -33,6 +33,7 @@ import org.eclipse.apoapsis.ortserver.components.pluginmanager.PluginManagerInte
 import org.eclipse.apoapsis.ortserver.components.pluginmanager.PluginOptionTemplate
 import org.eclipse.apoapsis.ortserver.components.pluginmanager.PluginOptionType
 import org.eclipse.apoapsis.ortserver.components.pluginmanager.PluginType
+import org.eclipse.apoapsis.ortserver.components.pluginmanager.mapToApi
 
 import org.ossreviewtoolkit.plugins.advisors.vulnerablecode.VulnerableCodeFactory
